requirements.txt
pandas
openpyxl
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
google-auth-httplib2
beautifulsoup4
groq
